+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Move loading of SQLite3Proxy to its own Railtie
 - Fix connection stickiness for inserts with Trilogy
 - Drop support for ActiveRecord 7.0.x
+- Drop support for Ruby 3.1.x
 
 ## [0.8.0] - 2025-08-28